Day 60
Day 60 of #dailycoding challenge ⬇️
The starting point for analyzing data in R is having access to the data. This data must be stored in one or more R objects. To use R, you must first know the different types of objects that can serve as "data containers" and know how to work with these objects. so today we will be interested in #Different #types #of #R #objects #serving #as #data #structure.
Objects whose elements are constrained to be all data of the same type are called "atomic", while the others are called "recursive".
*** atomic objects
1 vector: The vector is the simplest object, It has only one dimension.
2 matrix: The matrix is ​​a 2-dimensional generalization of the vector.
3 array: The last atomic object is the array, which generalizes the vector and the matrix to any number of dimensions.

*** recursive objects
1 list: A list is for a recursive object: it contains other objects, of any type.
2 data.farme: The data frame is a particular type of list, but it also shares characteristics with matrices, being deemed to have 2 dimensions (rows and columns). Its elements are objects constrained to be of corresponding dimensions. In the vast majority of cases, the elements are vectors or factors of the same length, forming the columns of the data frame.
3 factor: This type of object is used to store the observed values ​​of a categorical variable.

here the examples to understand more .⬇️
enjoy !!
#rstats #rladies #dailycoding #challenge