day1/season2
the starting point of programming is the creation of objects that will be manipulated later in our program. This is why the code of today consists of creating objects and controlling its modes.
The principal modes are "logical, numeric, complex and character".
we can convert the mode of an object to an other, but the conversion must respect this order :
logical
numeric
complex
character.
A logical mode object can be converted to any other mode, but for example a numeric mode object can only be converted to complex or character.